gvmap country_graph_coloring: fix passing in too short data for matrix element
ASan notices:

  ==44952==ERROR: AddressSanitizer: stack-buffer-overflow on address 0x7fff7de46374 at pc 0x7cc63f63a397 bp 0x7fff7de46280 sp 0x7fff7de45a28
  READ of size 8 at 0x7fff7de46374 thread T0
  #0 0x7cc63f63a396 in __interceptor_memcpy ../../../../src/libsanitizer/sanitizer_common/sanitizer_common_interceptors.inc:827
  #1 0x5a5144b45746 in SparseMatrix_coordinate_form_add_entry lib/sparse/SparseMatrix.c:1246
  #2 0x5a5144abe5c9 in country_graph_coloring cmd/gvmap/country_graph_coloring.c:115
  #3 0x5a5144a977de in map_optimal_coloring cmd/gvmap/make_map.c:103
  #4 0x5a5144a92e12 in makeMap cmd/gvmap/gvmap.c:354
  #5 0x5a5144a945a2 in mapFromGraph cmd/gvmap/gvmap.c:420
  #6 0x5a5144a94ac0 in main cmd/gvmap/gvmap.c:452
  #7 0x7cc63da29d8f in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
  #8 0x7cc63da29e3f in __libc_start_main_impl ../csu/libc-start.c:392
  #9 0x5a5144a8f064 in _start (/tmp/tmp.Bbb9M9dArO/bin/gvmap+0x16a064)

The sparse map abstraction uses the field `SparseMatrix_struct.type` to store
element type of a matrix and then pervasive `void *` parameters to offer generic
interfaces. A side effect of this is that bugs like that fixed in this change
(passing an `int *` instead of a `double *`) cannot be detected by the compiler.
This then results in reads out of bounds at runtime.

This bug was present in the very first revision of this code
(c193a91a0e9d8565e85bd6953a2c64e971d6ab79). Given this, combined with how easy
it is to trigger this bug with default settings, it is not clear to me how
`gvmap` can ever have worked.

Reported-by: Reported-by: GCC 11.4.0 -fsanitize=address

+++ cmd/gvmap/country_graph_coloring.c
@@ -112,7 +112,7 @@ void country_graph_coloring(int seed, SparseMatrix A, int **p) {
       const int jj = ja[j];
       if (jj != i){
 	nrow ++;
-	L = SparseMatrix_coordinate_form_add_entry(L, i, jj, &(int){-1});
+	L = SparseMatrix_coordinate_form_add_entry(L, i, jj, &(double){-1});
       }
     }
     L = SparseMatrix_coordinate_form_add_entry(L, i, i, &nrow);
